Refactor class responsible for scalling app window
