+++ client/src/main/java/com/battleships/Client.java
@@ -1,5 +1,6 @@
 package com.battleships;
 
+import com.battleships.windowScalling.WindowScalling;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
