+++ client/src/main/java/com/battleships/windowScalling/ScallingParams.java
@@ -1,0 +1,13 @@
+package com.battleships.windowScalling;
+
+class ScallingParams {
+    double width;
+    double height;
+    double ratio;
+
+    ScallingParams(double width, double height, double ratio) {
+        this.width = width;
+        this.height = height;
+        this.ratio = ratio;
+    }
+}
